+++ src/pools/pool_internal/test.rs
@@ -7,7 +7,6 @@ use std::time::Duration;
 
 use future::BoxFuture;
 use futures::prelude::*;
-use pretty_env_logger;
 use tokio::runtime::Runtime;
 use tokio::sync::oneshot;
 use tokio::time;
@@ -26,9 +25,7 @@ async fn check_out_fut<T: Poolable, M: Into<CheckoutConstraint>>(
     pool: &PoolInternal<T>,
     constraint: M,
 ) -> Result<Managed<T>, CheckoutError> {
-    pool.check_out(constraint)
-        .await
-        .map_err(|failure_package| failure_package.error_kind.into())
+    pool.check_out(constraint).await
 }
 
 #[test]